refactor: address PR #47 review comments
- NdpiService: drain stderr to logger at DEBUG level instead of discarding,
  so ndpiReader diagnostics are visible when troubleshooting
- NdpiService: pass full exception to log.warn() to include stack trace
- Frontend: extract shared getAppColor() utility (src/utils/appColors.ts)
  used by ConversationList, ConversationDetail, and AnalysisOverview,
  eliminating three copies of the hardcoded color fallback
