+++ frontend/src/pages/Analysis/AnalysisOverview.tsx
@@ -3,30 +3,12 @@ import { useOutletContext } from 'react-router-dom';
 import type { AnalysisData } from '@/types';
 import { AnalysisSummary } from '@components/analysis/AnalysisSummary';
 import { ProtocolBreakdownChart } from '@components/analysis/ProtocolBreakdown';
+import { getAppColor } from '@/utils/appColors';
 
 interface AnalysisOutletContext {
   data: AnalysisData;
   fileId: string;
 }
-
-const APP_COLORS: Record<string, string> = {
-  Zoom: '#2D8CFF',
-  WhatsApp: '#25D366',
-  Telegram: '#2AABEE',
-  Signal: '#3A76F0',
-  Discord: '#5865F2',
-  Teams: '#6264A7',
-  Skype: '#00AFF0',
-  Viber: '#7360F2',
-  WeChat: '#07C160',
-  YouTube: '#FF0000',
-  Netflix: '#E50914',
-  Spotify: '#1DB954',
-  TikTok: '#010101',
-  Instagram: '#E1306C',
-  Facebook: '#1877F2',
-  Twitter: '#1DA1F2',
-};
 
 export const AnalysisOverview = () => {
   const { data } = useOutletContext<AnalysisOutletContext>();
@@ -62,7 +44,7 @@ export const AnalysisOverview = () => {
                 key={app.name}
                 className="badge rounded-pill px-3 py-2 fs-6"
                 style={{
-                  backgroundColor: APP_COLORS[app.name] ?? '#6f42c1',
+                  backgroundColor: getAppColor(app.name),
                   color: '#fff',
                 }}
                 title={`${app.packets.toLocaleString()} packets · ${(app.bytes / 1024).toFixed(1)} KB`}
